Homework/site_project/lock_database/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.4
chardet==4.0.0
cryptography==3.3.1
defusedxml==0.7.0rc2
Django==3.1.3
django-sslserver==0.22
djangorestframework==3.12.1
idna==2.10
mysqlclient==2.0.3
oauthlib==3.1.0
pycparser==2.20
PyJWT==2.0.0
python3-openid==3.2.0
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
sqlparse==0.4.1
urllib3==1.26.2